+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/annotation/QNameAnnotator.kt
@@ -21,26 +21,27 @@ import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.intellij.lang.XPath
 import uk.co.reecedunn.intellij.plugin.intellij.lexer.XPathSyntaxHighlighterColors
-import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathEQName
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathNCName
 import uk.co.reecedunn.intellij.plugin.xpath.model.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xpath.lexer.IKeywordOrNCNameType
+import uk.co.reecedunn.intellij.plugin.xpath.model.XdmWildcardValue
 
 class QNameAnnotator : Annotator {
-    override fun annotate(element: PsiElement, holder: AnnotationHolder) {
-        if (element !is XPathEQName) return
-        if (element.language != XPath) return
+    override fun annotate(qname: PsiElement, holder: AnnotationHolder) {
+        if (qname !is XsQNameValue) return
+        if (qname.language != XPath) return
 
-        val qname = element as XsQNameValue
         val xmlns: Boolean
         if (qname.prefix != null) {
             if (qname.prefix!!.data == "xmlns") {
                 xmlns = true
-            } else {
+            } else if (qname.prefix !is XdmWildcardValue) {
                 xmlns = false
                 val prefix = qname.prefix?.element!!
                 holder.createInfoAnnotation(prefix, null).enforcedTextAttributes = TextAttributes.ERASE_MARKER
                 holder.createInfoAnnotation(prefix, null).textAttributes = XPathSyntaxHighlighterColors.NS_PREFIX
+            } else {
+                xmlns = false
             }
         } else {
             xmlns = false